+++ odoo/local-src/specific_discount_program/models/__init__.py
@@ -2,7 +2,9 @@
 # © 2016 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
+from . import account_invoice
 from . import partner_sponsor
+from . import product_template
 from . import program
 from . import program_condition
 from . import res_config