+++ odoo/local-src/specific_discount_program/models/product_template.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+# © 2016 Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from openerp import fields, models
+
+
+class ProductTemplate(models.Model):
+    _inherit = 'product.template'
+
+    no_quantity = fields.Boolean('No quantity')